Simplify symbols table selection for R/S
For Ruby and Sapphire, there are quite a lot of revisions (3 for English, 2 for the other languages.)

What we noticed, though, is that the symbol tables for `pokeruby_rev1.sym` and `pokeruby_rev2.sym` are identical (same for Sapphire), so there is no point in differentiating between them.

What's more, according to the [version history on Bulbapedia](https://bulbapedia.bulbagarden.net/wiki/Pok%C3%A9mon_Ruby_and_Sapphire_Versions#Version_history), it's likely that German/French/Italian/Spanish revision 0 is more similar to English revision 1 (which contains a bunch of internationalisation fixes). So it probably makes sense to use the symbol table for rev1 for these games. Anecdotally, this seems to have improved things a bit.

This simplifies the symbol table selection for R/S a bunch by just having these rules:

1. English Rev0 and Japanese use `pokeruby.sym`/`pokesapphire.sym`.
2. For German games, there is a symbols table straight from the decompliation project available, so they always use `pokeruby_de.sym`/`pokesapphire_de.sym`.
3. English Rev1, English Rev2, and all French, Italian, and Spanish games use `pokeruby_rev1.sym`/`pokesapphire_rev1.sym`.

This should make work on supporting the other languages a lot easier.