Debug Menu: Allow adding a Pokémon with an encounter-related ability
This adds some options to the debug menu that will add a Pokémon with certain encounter-related abilities (such as Illuminate, Sticky Hold, etc.) to the first party slot.

This Pokémon will be fainted, as weak as possible, and only knows moves that will make it faint.

Useful for testing encounter-boosting.

(On FR/LG and R/S only Illuminate is available, because all the other effects only got added in Emerald.)