Simplify Feebas mode pathfinding
This removes some special handling of land/surf transitions as well as going up the waterfall on Route 119, because that is now natively supported by our pathfinding.

It still really likes to go on land, only to start surfing again one or two tiles later.

That's because the pathfinding _really_ tries to avoid tiles with encounters on them. This could be easily fixed, but for some reason doing so introduces a loop.

More debugging is required to figure out why that is, so for now... good enough!