SGR: Check event var before allowing Hoenn Fossil resets
This adds a check for the in-game event var that controls the Hoenn fossil revival state before allowing the mode to start.

That prevents people from accidentally trying to start the mode before even giving a fossil to the scientist, or before the scientist is ready.

It also makes sure the player remembered to save.