Level Grind: Fix check whether party has damaging move
I've accidentally changed the condition for the Level Grind mode to start so it only looks at _non-fainted_ Pokémon to have a damaging move.

Meaning that if you had a Pokémon capable of causing damage but that Pokémon was fainted the mode would refuse to start -- even though we go heal before doing the grind anyway.