Do not crash if a profile with a GBC ROM exists
This adds handling for GB(C) ROMs.

Every time I switch between the gen2 branch and any other branch, I have to remove all the Gen2 profiles because otherwise the bot refuses to start due to the `*.gbc` files not being recognised as a valid ROM.

So this change adds support for these ROMs, but hides them. So as long as you don't try to _start_ a Gen2 profile, the mere existence of one won't cause an issue anymore.