Rock Smash: Exit Safari Zone when there aren't enough steps for another loops
This updates the Rock Smash mode on Emerald so that instead of just using up all their steps, the player will exit the Safari Zone using the Start Menu when there aren't enough steps left for another loop (that usually happens after the 3rd loop.)

This saves a few frames.

It also fixes a bug where the player wouldn't re-enter the Safari Zone if their steps ran out while still on the way to the Shuckle area.

And, finally, this adds a check that the player owns the Pokéblock Case -- which apparently is required to enter the SZ. Thanks to Neka for finding that out.