Stats: Only count catches if it's a shiny or CCF match
The stats system currently counts _every_ caught Pokémon of a given species, even if it's just the player manually catching something.

While not totally wrong as the field is just called `catches`, in practice we're using it to track how many shiny encounters have been successfully caught (in the stream overlay, for example.)

So it makes more sense to only count catches if it's an 'of-interest' Pokémon.