Improve cross-map pathing in Kanto Safari Zone
This updates the Safari mode so that it will use the direct connections between the central and northern/western areas if the player is able to use Surf. This saves some time by not having to traverse through the East and North areals.

I've also removed the custom warp code. If `navigate_to()` is used with a destination that is an arrow warp (i.e. needs another 'step' to activate the warp), the function will automatically follow that warp already.