Extract code for looking up the nearest Pokémon Center (#604)
This extracts the bulk of the code shared between EV Train and Level Grind modes: The lookup which Pokémon Centre is closest to a chosen training spot.

Having that function might be useful for other modes or features in the future, so separating it out seems sensible. Also, it avoids having two identical data lists (current map vs. potential nearest PCs) so there's no risk of someone updating one but not the other anymore.