EV Level: Stop if target is reached and verify map encounters (#605)
This adds a check to the EV Train mode to make it stop and switch to manual once the target EVs have been reached. Until now, it would just keep spinning and rejecting every battle. That made it a bit hard to figure out when the mode was actually done.

Also, it now checks the land encounter tables for the current map and verifies that at least one of the encounters will yield EVs that are useful towards the target.