Fix stale battle state being used in R/S Safari Zone when throwing Pokéblocks
When using a Pokéblock in R/S, after our block-throwing function has completed the game is already back in the action selection screen.

This means that the battle handler will never break out of that `while` loop (which is only meant to run for a single turn) and keeps providing the `decide_turn_in_safari_zone()` function with the battle state that has been generated before the Pokéblock has been thrown.

This fix just forces the battle handler to break out of the loop.