Prevent `on_battle_ended()` to run at the start of a battle
There is a brief moment after a battle commenced and before the battle handler callbacks are enabled, where all the conditions for 'we are no longer in a battle' are met.

That led to `on_battle_ended()` to be called at the start of a battle (as well as at the end.)

This fix just adds another check that the battle has actually started properly before considering to report the battle as done.