Fix pathfinding with Acro Bike and Mach Bike (#608)
This adds pathfinding support for traversing Muddy Slopes upwards using a Mach Bike.

Also, fixes two issues:

(1) When using the Acro Bike, the bot would sometimes try to access rails from the side due to a logic bug that skipped collision checking.

(2) If the bicycle wasn't registered to the `Select` button, using the `mount_bicycle()` utility would hang because it waits for the Start Menu to reappear after using the item. But that does not happen with bicycles.