Limit total number of save states in Load State Window (#616)
The Load State window crashes once a certain number of loadable states is reached. I haven't tested what that exact 'certain number' is.

We are already limiting the number of autosaves we display to 3.

But 'manual' save states (this includes those created with `Ctrl+S`, but also save states when encountering a shiny/CCF) are not filtered and/or limited.

This PR adds a limit of 128 total save states. That number has been chosen arbitrarily, but it doesn't crash my test profile anymore so I guess that's alright.