Pathfinding: Increase timeout when activating Surf
I've given the pathfinding algorithm 270 extra frames for waypoints that require activating Surf before triggering a pathing timeout for not having moved tiles for a while.

I don't remember how I chose that particular number, but it caused some issues for me on Emerald where the timeout got triggered before the bot was done starting to surf.

I've increased that limit to 300 frames and now it works. Oh well.