Fix timing of egg hatch reporting
Due to the egg hatch data structure being filled with some unrelated data, in the very first frame of the egg-hatching animation the EggHatchListener thought that the hatched Pokémon was already visible and would report the hatching immediately.

This fixes it so the hatched Pokémon is reported a couple of frames after the sprite becomes visible (about half a second later.)

I have updated the Static Gift Resets mode (Togepi, Wynaut) so it resets immediately when this happens rather than waiting for the game to get back to the overworld. This saves a few frames, too.