Fix bot sometimes hanging when trying to pathfind to a warp
In order to teach the bot to not 'accidentally' step on warping tiles, I have added an artifical cost to these tiles.

But that also means that if we specifically _tell_ the bot to pathfind towards a warp, the route would appear very expensive and the bot would keep looking for alternative ways that are cheaper.

Depending on the map and the situation, this might mean that the bot would try to check every conceivable path on the world map before accepting the route.