Rock Smash: Do not soft reset after catching a Pokémon
As Neka pointed out, the Rock Smash mode would just keep going after catching a Shiny/CCF Pokémon in Rock Smash mode -- which is an issue because these modes soft-reset after a while.

This is not an issue if `save_after_catching` is enabled because then soft-resetting would just load the game state _after_ catching the Pokémon.

So if that setting is NOT enabled (it's off by default), we switch to manual mode after catching something.