Add Item Steal mode
This mode will use the moves **Thief** or **Covet** to steal the held items from
wild Pokémon.

It works similarly to the Level Grind mode and supports the same locations.

On any supported route, stand on a tile that has encounters (tall grass or water)
and the mode will then spin to trigger encounters, use Thief/Covet once and run
away.

If the opponent was holding an item, that item will now be in the player's item
bag.

This can be useful to farm rare held items, or as an alternative to Pickup for
making money (by selling the stolen items afterwards.)