Verify that there is space in party/boxes before running modes that catch Pokémon
This adds a new assertion, `assert_boxes_or_party_can_fit_pokemon()`, which will verify that there is at least one empty slot in either the party or one of the boxes.

This assertion is used in each mode that is meant to find shinies (i.e. it's not checked in modes like Level Grind and EV Train -- even though they _could_ also encounter shinies that's not their primary purpose.)

I've also updated the `assert_player_has_poke_balls()` check in some modes that use soft resets to check the saved game and not the currently running one, as that's the actually important part.