Fix handling whiteout during Pokecenter Loop modes
When whiting out during a Pokécenter Loop (Level Grind, EV Train), the bot would run into an error while trying to access the first non-fainted Pokémon in the party -- of which there are none if the party has been defeated.

Also, the code assumed that the player would respawn _inside_ of the Pokémon Center, which is only true in FR/LG and R/S/E. So in the latter case, the bot would try to exit the building even though the player wasn't actually in it.

Fixes #651