Level Balancing Strategy: Go heal if weakest Pokémon is defeated (#655)
The Level Balancing battle strategy (used by the Level Grind mode) would not go healing until the entire party was unable to battle.

This meant that once the weakest Pokémon fainted, it would just no longer receive any EXP.

So it was quite possible that after a while, all the weak Pokémon in the party were defeated and only the strongest Pokémon got trained until it ran out of PP.

This changes it so that if the weakest Pokémon is defeated, the mode will go heal.