Level Grind: Go heal if lead Pokémon ran out of PP
In Level-balancing mode, once the lead Pokémon ran out of PP the bot would keep spinning for encounters but run away from them immediately.

That's because it was set up to only go heal once the entire party was unable to battle (since we're not just focussing on the lead) which wasn't the case, but it also wouldn't switch in other Pokémon.

On second thought, that didn't make all that much sense anyway. In Level-balancing mode, we want to get as much Exp. for the weakest Pokémon as possible, which is guaranteed to be the lead by the battle strategy's rotation logic.

So it makes sense to go and heal once the lead cannot fight anymore.

I've added an exception for the special case where the lead is not able to battle on its own anyway (Magikarp, Abra, ...) because in that case we want to switch in another Pokémon anyway.

This also fixes some bugs with `debug_create_pokemon()` that cause the `Edit Party` menu to (a) not set the remaining PP correctly, and (b) crash if creating a Pokémon with less than 4 moves.