Add a hidden whitelist file for using unsupported ROMs
This allows users to specify additional file hashes and/or file names that the bot should support.

That can be done by adding the hash and/or file name to `profiles/extra_allowed_roms.txt` (a file that does not exist by default.)

Doing so allows users to make minor modifications to ROMs as long as they don't change memory offsets.

There are several warnings, both in the Wiki and in the bot itself, that there will be no help when using unsupported files.