Track PokéNav calls in stats system
This updates the database schema, the stats system, and the PokéNav listener to store the number of PokéNav calls received during a shiny phase.

It also adds an optional event to the HTTP stream API so any API consumer can receive notifications about PokéNav calls.

Currently, this value is not really used anywhere. But it could be used for display in the stream overlay.