Stats: Fix invalid function call
Due to a silly typo the stats system tries to call `self.commit()` instead of `self._commit()` which is fixed by this PR.

Background: `commit()` needs to be called for database changes to actually be written. This was missing in some places, which meant that these changes were only written to disk once some other function actually called `commit()`.