Add menu option for clearing shiny phase stats
This introduces a new menu option for resetting the current shiny phase's stats to 0.

Can be useful to get rid of some random encounters that you do not really care about (individual encounters on the way to the current route or trainer battles, that would remain at 1 or 2 encounters because they don't exist in the place where the bot is now hunting.)

Or, like happened on stream, if the bot kept hunting even though we found our target already and only got moved after 1,500 more encounters.