Allow plugins and bot modes to prevent manual mode after hatched/gifted shiny
If the bot hatches or receives a shiny Pokémon (or a CCF match) it will currently always switch to manual mode.

This might not always be what we want: In daycare, we might need more than one shiny (for evolutions) and for gift soft-resets or starters, we might require a female shiny and therefore need to skip male ones.

This will be required for the stream plugin, so this new option is not being used by any default mode so far.