Fix keyboard navigation when trying to access the bottom row
The keyboard handler (used for the nicknaming feature) first tries to move the cursor to the correct row and only then to the correct column.

The problem is that the right-most row (with the 'Back' and 'OK' buttons) only has 3 buttons but the keyboard has 4 rows. The bottom-most button has the coordinates (8, 2) but if you press Down to reach the (non-existing) button (8, 3) it instead wraps back up to (8, 0).

So if the cursor is in that column and wants to get to the last row, it just gets stuck.

The solution was to just switch it around and navigate to the correct column first.