Fix Pokémon summary screen being detected as `TITLE_SCREEN`
The symbols table for Emerald uses `MainCB2` in several places: the title screen, the Pokémon summary screen, and the options menu.

So far if Callback2 was `MainCB2`, we set the game state to `TITLE_SCREEN` -- which never caused any issues, except I want to use it in the stream overlay.

This change renames the symbols so they all have unique names, and also introduces a new `POKEMON_SUMMARY_SCREEN` state.