Fix outdated state cache values being used after an emulator reset
I finally found the issue that triggered the bug in #687 and also what people have complained about on Discord:

In order to fix an issue with the stream overlay, I've added some code to the state cache so a value's age would never be negative (which would happen after a reset, because the frame number of when the data was taken would be higher than the current frame number.)

Instead, I let it report the age as 0 which turned out not to be a great idea. That's because in a lot of places we check whether the age is 0 and if so use the data from the cache. But in that case, that would be data from _before_ the reset.

So for example in the Starters mode, that would mean that the bot would check whether the party size is 1 by checking the party from before the reset, which would of course always be True.

Embarrassingly, this would have been easy to catch if I had bothered running the E2E test suite...