Use anti-shiny sprites in new overlay
I didn't know that, but apparently the old overlay used the anti-shiny sprite in the encounters list.

I guess I never saw that live, so I didn't implement that again.

Anyway, now that's back.

And also, for as long as an anti-shiny of a species is visible in the encounter log, it will appear as anti-shiny in the route encounters list.