Allow sending custom data via HTTP API, show PC storage numbers in overlay
This PR adds the ability for plugins to fill a `custom_state` dictionary that HTTP API clients can fetch (using the `/custom_state` endpoint.)

The idea being that plugins can send their own data to HTTP clients such as the overlay without having to modify the bot's code.

Similarly, there is a `CustomEvent` event type in the HTTP stream API for pushing data to clients.

Also, this adds an optional info bubble showing the fill state of the PC storage system/boxes.