Add 'release Pokémon from box' action for PC interaction (#700)
* Add 'release Pokémon from box' action for PC interaction

Allows bot modes to release Pokémon from a box in the `interact_with_pc()` function.

I'll need that for a mode that gets rid of all the filler Pokémon we're about to catch because of Castform.

* Use the 'Move Pokemon' menu for releasing Pokemon from boxes

That's because if the party is full, the 'Withdraw Pokemon' menu cannot be opened.