Fix glitching Box Fill State info bubble in overlay
Thanks to a missing assignment, the timer interval for updating the box fill state bubble would not only be initialised once, but instead a new timer would be created each time the info bubbles were updated -- which happens once for every wild encounter.

So after a few thousand encounters, you'd end up with thousands of timers all trying to update the box bubble. Which leads to visual glitching.