Fix some message-handling listeners stopping prematurely (#706)
We have a bunch of listeners that wait for some event (such as a PokeNav call, or a Pokémon fainting in the overworld due to poison, or whiting out, or Safari Zone steps being used up) and then handle it by adding a handling function to the controller stack.

This works fine in most cases, but sometimes -- and I haven't figured out why, but it seems to be related to battle-handling code -- that controller gets removed before it finishes handling the message.

That's the reason why we got that sandstorm staring contest in the desert, for example.

This is more of a 'dealing with the symptoms' kind of fix, but anyway.

Instead of using an internal flag to check whether the handling function is active, it _actually_ checks whether the handling function is part of the controller stack.

With that, PokeNav handling was perfectly reliable for me.

I also added the 'Force PokeNav' debug tool in the debug menu that I used for testing.