Fix Feebas mode starting to search for the tile again when the in-game day changes
The Feebas mode is programmed on the assumption that the Feebas tiles change once the in-game clock changes to a new day.

That is, apparently, not the case. So this PR removes that behaviour and replaces it with a safeguard that will restart the search for a Feebas tile if, for whatever reason, it has not seen a Feebas within 20 encounters.

It also fixes the overlay to show the correct encounter rates after a Feebas tile has been found.