Fix issues with Surfing during pathfinding (#716)
When using the pathfinding utilities, if the player avatar was on a water tile (i.e. surfing) at the start of the path, this would disable running for the entirety of the path.

The reason was that 'running' involves pressing `B`, which on a deep sea tile would lead to the player diving.

That's been fixed now. The bot will not try to run while surfing, but start as soon as the player avatar reaches land.

Also, this fixes a bug where the pathfinding algorithm wouldn't take into account dynamic obstacles (NPCs, overworld items, etc.) during a transition from water to land.