Support fishing in PokecenterLoop utility (#717)
The PokecenterLoop utility is used internally for modes like Level Grind and EV-train.

So far, the utility only supported spinning but there are cases where fishing might be useful to (for example when farming speed EVs without gaining a lot of EXP, Magikarp is a good target)

So this adds a switch to that mode. The bot itself doesn't use it (yet)