Fix pathfinding in Shoal Cave
There are a few 'hidden' items in Shoal Cave that appear as little mounds in the cave.

If the item hasn't been picked up, these are not traversable.

But the game doesn't add them as entities (as is the case with 'item balls' in the overworld) but instead dynamically modifies the map itself. Due to caching the map data, our pathfinding never picks up on that.

So I've added a list of hardcoded coordinates to check for, which fixes the issue.