Fix crash in GenerateEncounterMedia plugin
If the bot is started and the save state is _during_ a shiny encounter, then for the next encounter the `GenerateEncounterMedia` plugin will raise an error because it cannot find its listener in the controller stack even though it has seen a shiny encounter.

That usually only happened during development, but it was mildly annoying.

Also, this PR fixes a few type hints.