Prevent notification for duplicate encounters
Previously, if you loaded one of the save states that is automatically created when encountering a shiny, the bot would immediately send a desktop notification and create a save state because the battle/egg listener would see it as a new encounter.

I've fixed this so it at least checks whether the encounter equals the previously last-seen encounter. So this won't help if you load an older save state, but at least the typical testing scenario of restarting the bot and then ending up with the auto save is covered.

In addition to that, I have added a (hidden) config option: Setting `create_save_state_for_shiny` to `false` in `logging.yml` will disable this automatic creation of save states entirely. Useful for testing when you cheat in a lot of shinies.