Faster berry blending (#749)
* Faster berry blending.

Uses a wider hit range since hits at the start increase speed in
addition to perfects.

Alternates between A and L presses, since with the L=A setting you can
do a full A press on every frame, including hitting 2 perfects on some
early passes.

Intentionally misses at the end of the blend to slow the blender down,
which brings up the results screen faster.  It uses the memory which
diplays the red bar at the top to know when the blend is close to done.

* Update wiki to mention L=A

* Update progress value for braking to work for 2 player blends.