Overlay: Fix reload loop in connection checking code (#754)
The overlay listened to the wrong event to see if it received any messages. (Turns out, the `message` event is a special one and is not triggered for custom events.) That meant that after 15 seconds, the overlay would assume that it hadn't received anything and would attempt to reconnect. And it would do so once a second which causes significant lag.

A second issue was that after noticing that the connection may have been lost, it would run `doFullUpdate()` which now has auto-retry behaviour. So for each second that the connection was lost, it would run _another_ instance of `doFullUpdate()` which would quickly lead to a massive list of concurrent requests. And once the connection was re-established, each instance would then start its own event stream, meaning that each event would be handled multiple times.