Overlay: Do not show '1 missed' during a shiny encounter
Because the overlay would already increase `shiny_encounters` at the start of a battle, it would show '1 missed' in the route encounters table because the number of catches is lower than the number of encounters.

I've changed it so it doesn't increase the `shiny_encounters` value. The overlay fetches the full stats from the HTTP API after a shiny encounter ended, thus fixing the encounter numbers.