Rock Smash: Lower the risk of encountering duplicate PVs in Safari Zone
In the Safari Zone (but probably also in Granite Cave) we see elevated levels of duplicate encounters.

That's presumably because it takes to long to get encounters that it's fairly likely to end up with the same RNG sequence after a while, just due to different random overworld events.

This PR tries to mitigate that by (a) adding single-frame delays whenever a duplicate has been detected, and (b) removing the limit of how much cash the bot will spend in the Safari Zone before soft-resetting. It will now use up all the player's cash before resetting, rather than after a flat 50 SZ entries.