Fix desktop notification and PK3 creation
The duplicate detection is running in two places: `log_encounter()` and `handle_encounter()`.

Because `log_encounter()` also _sets_ the last encounter in the stats DB, the subsequent check in `handle_encounter()` will _always_ mark the encounter as duplicate and thus not create a desktop notification, PK3 file, or even save state.

This PR changes it so that (a) the duplicate detection is only run once while creating the `EncounterInfo` instance, and (b) updating the Starters mode so it does not create a _second_ instance of `EncounterInfo` on R/S/E which would cause the same issue.